+++ lib/hooks/use-custom-theme.ts
@@ -1,11 +1,15 @@
 import { useEffect } from 'react';
-import { setGlobalColor } from '../helpers/styles';
+import { setGlobalStyle, getHSLFromHex } from '../helpers/styles';
 import type { Theme } from '../Chat';
 
 export default function useCustomTheme (theme?: Theme) {
   useEffect(() => {
     if (theme?.primaryColor) {
-      setGlobalColor('--enhancedocs-primary-base', theme.primaryColor);
+      const { primaryColor } = theme;
+      const hsl = getHSLFromHex(primaryColor);
+
+      setGlobalStyle('--base-color', `${hsl.hue}, ${hsl.saturation}%`);
+      setGlobalStyle('--enhancedocs-primary-base', primaryColor);
     }
   }, []);
 }